Fix RingModulatorGainAdjust type

+++ TAC_COM/Models/EffectParameters.cs
@@ -15,7 +15,7 @@ namespace TAC_COM.Models
         public float DistortionDry = 0.5f;
 
         public Type? RingModulatorType;
-        public int RingModulatorGainAdjust = 0;
+        public float RingModulatorGainAdjust = 0;
         public Dictionary<string, object> RingModulatorParameters = [];
 
         public float HighpassFrequency;
